vision/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
# Image processing
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
# ML models (depth estimation + segmentation) — heavy, not loaded until worker starts
torch==2.3.0
torchvision==0.18.0
transformers==4.41.1
# AWS S3
boto3==1.34.98
# Job queue (Redis)
redis==5.0.4
# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
